+++ src/OpenTelemetry/CHANGELOG.md
@@ -6,6 +6,10 @@ Notes](../../RELEASENOTES.md).
 
 ## Unreleased
 
+* Fixed CA1815 by implementing `IEquatable<T>`, `operator ==`, and `operator !=`
+  on `HistogramBucket` and `LogRecordScope`.
+  ([#6278](https://github.com/open-telemetry/opentelemetry-dotnet/issues/6278))
+
 * Fix incorrect validation of `OTEL_BSP_*` and `OTEL_BLRP_*` environment
   variables.
   ([#7187](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7187))